AK+LibHTTP: Represent bidirectional async streams as pairs of ends
... so that we always treat bidirectional connections as two independent
streams.

While representing duplex streams as classes inheriting both from
AsyncInputStream and AsyncOutputStream is convenient and natural, having
a single shared AsyncResource backing both of them will significantly
complicate code in situations when one treats ends as independent (and
this is pretty much any code that would have casted AsyncStream to one
of its bases).

Additionally, with this commit, there (completely accidentally) emerges
a nice meaning of *Connection suffix -- as something dealing with a pair
of input and output streams.