AK+LibHTTP+LibTest: Make AsyncResource::reset almost synchronous
This commit fixes a fundamental flaw with the previous implementation of
cancellation / closing of resources. Suppose an asynchronous resource
spins up a background coroutine that exists and does some work even when
nobody awaits the resource (read: AsyncStreamTransform starts generator
that tries to read from the underlying stream in between enqueue_some
iterations of the transform itself) and a request to cancel the
asynchronous operation arrives.

Previously, we would have had to synchronously reset the resource and
prepare it for the destruction. But this required cancelling the
background coroutine synchronously! While this can be done in theory,
maintaining enough information in coroutine frames increases the memory
footprint of each frame and, more importantly, AK::Coroutine was never
designed to support cancellation on arbitrary suspension points.
Therefore, the design of AsyncResource effectively made use of
background coroutines impossible.

Instead of that, here we introduce the notion of "almost synchronous"
execution (returning control to the caller in the same event loop
iteration) and require AsyncResource::reset to be almost synchronous
instead of just being synchronous. This elegantly fixes the issue with
background coroutines: if they are cooperating, it is trivial to join
them in an almost synchronous fashion.

Unfortunately, destructors cannot be made asynchronous, so automatic-reset-on-destruction feature is deleted. This makes us fight
against the C++ type system in more places but note that we did that
already with
```
if (result.is_error()) {
    m_stream.reset();
    co_return result.release_error();
}
```
construct. Additionally, this can be usually combated with
`if (m_stream.is_open()) co_await m_stream.reset();` checks placed
strategically to alleviate biggest pain points.

While on it, the commit also fixes a long-standing design error with
concurrent awaiting and resetting the resource. Practice shows that
concurrent asynchronous operations on the same resource lead to
unreadable mess of synchronization flags that is almost impossible to
write and, later, maintain. Instead, we ban concurrent resets altogether
and introduce synchronous HybridResource::cancel with clear and simple
semantics of cancelling the current awaiter with an error or scheduling
an error for the next attempt to wait on the resource. This allows to
reuse the existing "if method returned error, just destroy resource"
logic we already have for handling I/O errors for cancellation in
callers.